feat: for first review 14 Project Work

+++ controllers/cards.js
@@ -1,51 +1,55 @@
 const Card = require('../models/Card');
+const { Types } = require('mongoose');
 const {
   DataError,
   NotFoundError,
-  serviceError,
+  ForbiddenError,
 } = require('../utils/Errors');
 
-const createCard = (req, res) => {
+const createCard = (req, res, next) => {
   const { name, link } = req.body;
   return Card.create({ name, link, owner: req.user._id })
     .then((card) => res.status(201).send({ data: card }))
     .catch((err) => {
       if (err.name === 'ValidationError') {
-        return res.status(400).send({ message: `Переданы некорректные данные: ${err.message}` });
+        next(new DataError(`Переданы некорректные данные: ${err.message}`))
       }
-      return res.status(serviceError.statusCode).send({ message: serviceError.message + err.message });
+      next(err);
     });
 };
 
-const getAllCards = (req, res) => {
+const getAllCards = (req, res, next) => {
   return Card.find({})
     .then((cards) => res.status(200).send({ data: cards }))
-    .catch((err) => res.status(500).send({ message: `Somethin wrong with server: ${err.message}` }));
+    .catch(next);
 };
 
-const deleteCard = async (req, res) => {
+const deleteCard = async (req, res, next) => {
   try {
     const { cardId } = req.params;
+    const card = await Card.findById(cardId);
+    if (!card) {
+      throw new NotFoundError(`Карточка с идентификатором ${cardId} не была найдена и не удалена.`);
+    }
+    if (card.owner._id.toString() !== req.user._id) {
+      throw new ForbiddenError('У Вас недостаточно прав чтобы удалить эту карточку.');
+    }
     const cardById = await Card.findByIdAndRemove(cardId);
     if (cardById) {
       return res.status(200).send({ data: cardById });
     }
-    throw new NotFoundError(`Карточка с идентификатором ${cardId} не была найдена и не удалена.`);
   } catch (err) {
-    if (err instanceof NotFoundError) {
-      return res.status(err.statusCode).send({ message: err.message });
+    if (err.name === 'CastError') {
+      next(new NotFoundError(`Карточка с идентификатором ${req.params.cardId} не была найден и не удалена.`));
     }
-    return res.status(serviceError.statusCode).send({ message: serviceError.message + err.message });
+    next(err);
   }
 };
 
-const putLike = async (req, res) => {
+const putLike = async (req, res, next) => {
   try {
     const { cardId } = req.params;
     const unupdatedCard = await Card.findByIdAndUpdate(cardId, { $addToSet: { likes: req.user._id } });
-    if (!unupdatedCard) {
-      throw new NotFoundError(`Не удалось найти карточку с идентификатором ${cardId} чтобы поставить лайк.`);
-    }
     if (unupdatedCard.likes.includes(req.user._id)) {
       throw new DataError('Данный пользователь уже ставил лайк этой карточке.');
     } else {
@@ -53,23 +57,14 @@ const putLike = async (req, res) => {
       return res.status(200).send({ data: updatedCard });
     }
   } catch (err) { // все случаи ошибок из описания здесь описаны
-    if (err instanceof DataError) { // на случай если пользователь уже ставил лайк. 400
-      return res.status(err.statusCode).send({ message: err.message });
-    }
-    if (err instanceof NotFoundError) { // на случай если _id карточки не валидный. 404
-      return res.status(err.statusCode).send({ message: err.message });
-    }
-    return res.status(serviceError.statusCode).send({ message: serviceError.message + err.message }); // 500
+    next(err);
   }
 };
 
-const removeLike = async (req, res) => {
+const removeLike = async (req, res, next) => {
   try {
     const { cardId } = req.params;
     const unupdatedCard = await Card.findByIdAndUpdate(cardId, { $pull: { likes: req.user._id } });
-    if (!unupdatedCard) {
-      throw new NotFoundError(`Не удалось найти карточку с идентификатором ${cardId} чтобы снять лайк.`);
-    }
     if (!unupdatedCard.likes.includes(req.user._id)) {
       throw new DataError('Данный пользователь не ставил лайк этой карточке.');
     } else {
@@ -77,13 +72,7 @@ const removeLike = async (req, res) => {
       return res.status(200).send({ data: updatedCard });
     }
   } catch (err) { // все случаи ошибок из описания здесь описаны
-    if (err instanceof DataError) { // на случай если пользователь не ставил лайк. 400
-      return res.status(err.statusCode).send({ message: err.message });
-    }
-    if (err instanceof NotFoundError) { // на случай если _idкарточки не валидный. 404
-      return res.status(err.statusCode).send({ message: err.message });
-    }
-    return res.status(serviceError.statusCode).send({ message: serviceError.message + err.message }); // 500
+    next(err);
   }
 };
